// src/services/rest-api/customer-service/model/fatca-w8-input.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FatcaW8Input
 */
export interface FatcaW8Input {
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'citizenCountryCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'address1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'address2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'address3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'countryCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'mailingAddress1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'mailingAddress2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'mailingAddress3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'mailingCountryCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'taxNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'foreignTaxNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'foreignTaxNoRemark'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'referenceNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'dateOfBirth'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'beneficialTaxRate'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'additionCondition'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'beneficialTaxRemark1'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'beneficialTaxRemark2'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'beneficialTaxRemark3'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'beneficialCountryCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'name'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'refId'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof FatcaW8Input
     */
    'refType'?: string | null;
}